requirements.txt
flask
flask-cors
numpy
scikit-learn
google-genai
python-dotenv
gunicorn
